goshawk-api/requirements.txt
Django
djangorestframework
markdown
django-filter
drf-yasg